Refactor CI scripts to use the updated docker images (with gpadmin user already present) (#50)

+++ concourse/scripts/pxf_common.bash
@@ -1,6 +1,6 @@
 #!/bin/bash -l
 
-if [ "${TARGET_OS}" == "ubuntu" ]; then
+if [[ "${TARGET_OS}" == "ubuntu" ]]; then
 	GPHOME="/usr/local/gpdb"
 else
 	GPHOME="/usr/local/greenplum-db-devel"
@@ -15,7 +15,7 @@ PXF_CONF_DIR="/home/gpadmin/pxf"
 
 JAVA_HOME=$(ls -d /usr/lib/jvm/java-1.8.0-openjdk* | head -1)
 
-if [ -d gpdb_src/gpAux/extensions/pxf ]; then
+if [[ -d gpdb_src/gpAux/extensions/pxf ]]; then
 	PXF_EXTENSIONS_DIR=gpdb_src/gpAux/extensions/pxf
 else
 	PXF_EXTENSIONS_DIR=gpdb_src/gpcontrib/pxf
@@ -50,22 +50,22 @@ function run_regression_test() {
 
 function install_gpdb_binary() {
 
-	if [ ! -d ${GPHOME}/bin/psql ]; then
+	if [[ ! -d ${GPHOME}/bin/psql ]]; then
 		mkdir -p ${GPHOME}
 		tar -xzf bin_gpdb/*.tar.gz -C ${GPHOME}
 	fi
 
-	if [ "${TARGET_OS}" == "centos" ]; then
+	if [[ "${TARGET_OS}" == "centos" ]]; then
 		service sshd start
 		psi_dir=$(find /usr/lib64 -name psi | sort -r | head -1)
-	elif [ "${TARGET_OS}" == "ubuntu" ]; then
+	elif [[ "${TARGET_OS}" == "ubuntu" ]]; then
 		service ssh start
 		pip install psi
 		psi_dir=$(find /usr/local/lib -name psi | sort -r | head -1)
 	fi
 
 	# Copy PSI package from system python to GPDB as automation test requires it
-	if [ ! -d ${GPHOME}/lib/python/psi ]; then
+	if [[ ! -d ${GPHOME}/lib/python/psi ]]; then
 		cp -r ${psi_dir} ${GPHOME}/lib/python
 	fi
 
@@ -85,12 +85,12 @@ function remote_access_to_gpdb() {
 	cp cluster_env_files/.ssh/*.pem /home/gpadmin/.ssh/id_rsa
 	cp cluster_env_files/public_key.openssh /home/gpadmin/.ssh/authorized_keys
 	{ ssh-keyscan localhost; ssh-keyscan 0.0.0.0; } >> /home/gpadmin/.ssh/known_hosts
-	ssh ${SSH_OPTS} gpadmin@mdw "source ${GPHOME}/greenplum_path.sh && \
-	  export MASTER_DATA_DIRECTORY=${MDD_VALUE} && \
-	  echo 'host all all 10.0.0.0/16 trust' >> ${MDD_VALUE}/pg_hba.conf && \
-	  psql -d template1 -c 'CREATE EXTENSION pxf;' && \
-	  psql -d template1 -c 'CREATE DATABASE gpadmin;' && \
-	  psql -d template1 -c 'CREATE ROLE root LOGIN;' && \
+	ssh ${SSH_OPTS} gpadmin@mdw "source ${GPHOME}/greenplum_path.sh &&
+	  export MASTER_DATA_DIRECTORY=${MDD_VALUE} &&
+	  echo 'host all all 10.0.0.0/16 trust' >> ${MDD_VALUE}/pg_hba.conf &&
+	  psql -d template1 -c 'CREATE EXTENSION pxf;' &&
+	  psql -d template1 -c 'CREATE DATABASE gpadmin;' &&
+	  psql -d template1 -c 'CREATE ROLE root LOGIN;' &&
 	  gpstop -u"
 }
 
@@ -116,28 +116,36 @@ function add_remote_user_access_for_gpdb() {
 
 function setup_gpadmin_user() {
 
-	groupadd -g 1000 gpadmin && useradd -u 1000 -g 1000 -M gpadmin
-	echo "gpadmin  ALL=(ALL)       NOPASSWD: ALL" > /etc/sudoers.d/gpadmin
-	groupadd supergroup && usermod -a -G supergroup gpadmin
-	mkdir -p /home/gpadmin/.ssh
-	ssh-keygen -t rsa -N "" -f /home/gpadmin/.ssh/id_rsa
-	cat /home/gpadmin/.ssh/id_rsa.pub >> /home/gpadmin/.ssh/authorized_keys
-	chmod 0600 /home/gpadmin/.ssh/authorized_keys
-	{ ssh-keyscan localhost; ssh-keyscan 0.0.0.0; } >> /home/gpadmin/.ssh/known_hosts
-	chown -R gpadmin:gpadmin ${GPHOME} /home/gpadmin
-	echo -e "password\npassword" | passwd gpadmin 2> /dev/null
+    if [[ "${TARGET_OS}" == "ubuntu" ]]; then
+        ./gpdb_src/concourse/scripts/setup_gpadmin_user.bash
+    else
+        # Don't create gpadmin user if already exists
+        gpadmin_exists=$(getent passwd gpadmin || true | wc -l)
+        if [[ ${gpadmin_exists} == 0 ]]; then
+            groupadd -g 1000 gpadmin && useradd -u 1000 -g 1000 -M gpadmin
+            echo "gpadmin  ALL=(ALL)       NOPASSWD: ALL" > /etc/sudoers.d/gpadmin
+            groupadd supergroup && usermod -a -G supergroup gpadmin
+            mkdir -p /home/gpadmin/.ssh
+            ssh-keygen -t rsa -N "" -f /home/gpadmin/.ssh/id_rsa
+            cat /home/gpadmin/.ssh/id_rsa.pub >> /home/gpadmin/.ssh/authorized_keys
+            chmod 0600 /home/gpadmin/.ssh/authorized_keys
+            { ssh-keyscan localhost; ssh-keyscan 0.0.0.0; } >> /home/gpadmin/.ssh/known_hosts
+            chown -R gpadmin:gpadmin ${GPHOME} /home/gpadmin
+            echo -e "password\npassword" | passwd gpadmin 2> /dev/null
+        fi
+    fi
 	echo -e "gpadmin soft core unlimited" >> /etc/security/limits.d/gpadmin-limits.conf
 	echo -e "gpadmin soft nproc 131072" >> /etc/security/limits.d/gpadmin-limits.conf
 	echo -e "gpadmin soft nofile 65536" >> /etc/security/limits.d/gpadmin-limits.conf
 	echo -e "export JAVA_HOME=${JAVA_HOME}" >> /home/gpadmin/.bashrc
-	if [ -d gpdb_src/gpAux/gpdemo ]; then
+	if [[ -d gpdb_src/gpAux/gpdemo ]]; then
 		chown -R gpadmin:gpadmin gpdb_src/gpAux/gpdemo
 	fi
 }
 
 function install_pxf_client() {
 	# recompile pxf.so file for dev environments only
-	if [ "${TEST_ENV}" == "dev" ]; then
+	if [[ "${TEST_ENV}" == "dev" ]]; then
 		source ${GPHOME}/greenplum_path.sh
 		source /opt/gcc_env.sh || true
 
@@ -204,13 +212,14 @@ function setup_impersonation() {
 EOF
 		sed -i -e '/<configuration>/r proxy-config.xml' ${GPHD_ROOT}/hadoop/etc/hadoop/core-site.xml ${GPHD_ROOT}/hbase/conf/hbase-site.xml
 		rm proxy-config.xml
-	elif [ "${IMPERSONATION}" == "false" ]; then
-		echo 'Impersonation is disabled, no proxy user setup performed.'
+	elif [[ "${IMPERSONATION}" == "false" ]]; then
+		echo "Impersonation is disabled, no proxy user setup performed."
 	else
 		echo "ERROR: Invalid or missing CI property value: IMPERSONATION=${IMPERSONATION}"
 		exit 1
 	fi
-	if [ ! -f ${GPHD_ROOT}/hbase/lib/pxf-hbase-*.jar ]; then
+    pxf_hbase_jar=$(find ${GPHD_ROOT}/hbase/lib -name pxf-hbase-*.jar | wc -l)
+	if [[ "$pxf_hbase_jar" == "0" ]]; then
 		cp ${PXF_HOME}/lib/pxf-hbase-*.jar ${GPHD_ROOT}/hbase/lib
 	fi
 }
@@ -231,7 +240,7 @@ function start_hadoop_services() {
 	jps
 
 	# grant gpadmin user admin privilege for feature tests to be able to run on secured cluster
-	if [ "${IMPERSONATION}" == "true" ]; then
+	if [[ "${IMPERSONATION}" == "true" ]]; then
 		echo 'Granting gpadmin user admin privileges for HBase'
 		echo "grant 'gpadmin', 'RWXCA'" | hbase shell
 	fi
@@ -244,7 +253,7 @@ function init_and_configure_pxf_server() {
 	su gpadmin -c "PXF_CONF=${PXF_CONF_DIR} ./bin/pxf init"
 
 	# update impersonation value based on CI parameter
-	if [ ! "${IMPERSONATION}" == "true" ]; then
+	if [[ ! "${IMPERSONATION}" == "true" ]]; then
 		echo 'Impersonation is disabled, updating pxf-env.sh property'
 		su gpadmin -c "echo 'export PXF_USER_IMPERSONATION=false' >> ${PXF_CONF_DIR}/conf/pxf-env.sh"
 	fi
